\name{begin-methods}
\docType{methods}
\alias{begin-methods}
\alias{begin}
\title{Return to the beginning of the file.}
\description{
  Sets the filepointer to the beginning of the file. The next call to 
  \code{next_block} returns the first lines of the file. This method is usually
  used in combination with \code{next_block} 
}
\section{Methods}{
  \describe{
    \item{\code{signature(x = "laf")}}{
      See \code{\linkS4class{laf}}. 
    }
  }
}
\keyword{methods}

