% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf.R
\docType{methods}
\name{close,laf-method}
\alias{close,laf-method}
\title{Close the connection to the Large File}
\usage{
\S4method{close}{laf}(con, ...)
}
\arguments{
\item{con}{a \code{"\link[=laf-class]{laf}"} object that can be closed.}

\item{...}{unused.}
}
\description{
Close the connection to the Large File
}

