\name{current_line-methods}
\docType{methods}
\alias{current_line-methods}
\alias{current_line}
\title{Get the current line in the file.}
\description{
  Returns the next line that will be read by \code{next_block}. The current line
  can be set by the method \code{goto}.
}
\section{Methods}{
  \describe{
    \item{\code{signature(x = "laf")}}{
      See \code{\linkS4class{laf}}. 
    }
  }
}
\keyword{methods}

