\name{goto-methods}
\docType{methods}
\alias{goto-methods}
\alias{goto}
\title{Go to specified line in the file}
\description{
  Sets the current line to the line number specified. The next call to
  \code{next_block} will return the data on the specified line in the first row.
  The current line can be read by \code{current_line}.
}
\section{Methods}{
  \describe{
    \item{\code{signature(x = "laf", i = "numeric")}}{
      See \code{\linkS4class{laf}}. 
    }
  }
}
\keyword{methods}

