\docType{class}
\name{laf_column-class}
\alias{laf_column-class}
\title{Column of a Large File Object}
\description{
Representation of a column in a Large File object. This
class itself is a subclass of the class \code{laf}. In
principle all methods that can be used with a \code{laf}
object can also be used with a \code{laf_column} object
except the the \code{column} or \code{columns} arguments of
these methods are not needed.
}
\section{Objects from the Class}{
  Object of this class are usually created by using the
  \code{$} operator on \code{laf} objects.
}

