\docType{methods}
\name{names,laf-method}
\alias{names,laf-method}
\alias{names<-,laf-method}
\title{Get the names of the columns in a Large File object}
\usage{
\S4method{names}{laf}(x)

\S4method{names}{laf}(x) <- value
}
\arguments{
  \item{x}{an \R object.}

  \item{value}{a character vector of up to the same length
  as \code{x}, or \code{NULL}.}
}
\description{
Get the names of the columns in a Large File object
}

