% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf.R
\docType{methods}
\name{nrow,laf-method}
\alias{nrow,laf-method}
\title{Get the number of rows in a Large File object}
\usage{
\S4method{nrow}{laf}(x)
}
\arguments{
\item{x}{a \code{"\link[=laf-class]{laf}"} object.}
}
\description{
Get the number of rows in a Large File object
}
