\docType{methods}
\name{show,laf_column-method}
\alias{show,laf-method}
\alias{show,laf_column-method}
\title{Print a column of a Large File object to screen}
\usage{
\S4method{show}{laf_column}(object)

\S4method{show}{laf}(object)
}
\arguments{
  \item{object}{the object to print to screen.}
}
\description{
Print a column of a Large File object to screen

Print the Large File object to screen
}

