\docType{methods}
\name{close,laf-method}
\alias{close,laf-method}
\title{Close the connection to the Large File}
\usage{
\S4method{close}{laf}(con, ...)
}
\arguments{
  \item{con}{a connection.}

  \item{...}{arguments passed to or from other methods.}
}
\description{
Close the connection to the Large File
}

