% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf.R, R/laf_column.R
\docType{methods}
\name{levels,laf-method}
\alias{levels,laf-method}
\alias{levels,laf_column-method}
\alias{levels<-,laf-method}
\alias{levels<-,laf_column-method}
\title{Get and change the levels of the column in a Large File object}
\usage{
\S4method{levels}{laf}(x)

\S4method{levels}{laf}(x) <- value

\S4method{levels}{laf_column}(x)

\S4method{levels}{laf_column}(x) <- value
}
\arguments{
\item{x}{a \code{"\link[=laf-class]{laf}"} object.}

\item{value}{a list with the levels for each column.}
}
\description{
Get and change the levels of the column in a Large File object
}

