% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf.R
\name{names,laf-method}
\alias{names,laf-method}
\alias{names<-,laf-method}
\title{Get and set the names of the columns in a Large File object}
\usage{
\S4method{names}{laf}(x)

\S4method{names}{laf}(x) <- value
}
\arguments{
\item{x}{a \code{"\link[=laf-class]{laf}"} object.}

\item{value}{a character vector with the new column names}
}
\description{
Get and set the names of the columns in a Large File object
}
