% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laf.R, R/laf_column.R
\docType{methods}
\name{show,laf-method}
\alias{show,laf-method}
\alias{show,laf_column-method}
\title{Print the Large File object to screen}
\usage{
\S4method{show}{laf}(object)

\S4method{show}{laf_column}(object)
}
\arguments{
\item{object}{the object to print to screen.}
}
\description{
Print the Large File object to screen

Print a column of a Large File object to screen
}
